shape: code hygiene, move packages
