+++ src/main/scala/org/zouzias/spark/lucenerdd/spatial/shape/context/ContextLoader.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zouzias.spark.lucenerdd.spatial
+package org.zouzias.spark.lucenerdd.spatial.shape.context
 
 import java.io.{StringReader, StringWriter}
