+++ src/test/scala/org/zouzias/spark/lucenerdd/spatial/shape/ShapeLuceneRDDSpec.scala
@@ -21,9 +21,9 @@ import java.io.StringWriter
 import com.holdenkarau.spark.testing.SharedSparkContext
 import com.spatial4j.core.distance.DistanceUtils
 import org.scalatest.{BeforeAndAfterEach, FlatSpec, Matchers}
-import org.zouzias.spark.lucenerdd.spatial.ContextLoader
 import org.zouzias.spark.lucenerdd.testing.LuceneRDDTestUtils
 import org.zouzias.spark.lucenerdd._
+import org.zouzias.spark.lucenerdd.spatial.shape.context.ContextLoader
 
 class ShapeLuceneRDDSpec extends FlatSpec
   with Matchers
